backend/requirements.txt
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.22
protobuf==3.14.0
PyJWT==2.0.1
pylint==2.6.0
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
